Allows our console widget to handle commands.
Adds a `handle_message` function to the console that allows executing
commands entered by the user.

Also adds a `/reverse_console` that reverses the way logs are displayed.
Instead of the latest log being at the bottom, it'll be at the top.
I added this command because my screen resolution is kinda crappy and the
last logs weren't displaying correctly.